+++ playtime/src/controllers/dashboard-controller.js
@@ -13,6 +13,23 @@ export const index = async (request, h) => {
   });
 };
 
+export const addPlaylistFailAction = async (request, h, err) => {
+  const user = request.auth.credentials;
+  const userName = user ? `${user.firstName} ${user.lastName}`.trim() : '';
+  const playlists = await playlistStore.getUserPlaylists(user._id);
+  return h
+    .view('dashboard-view', {
+      title: 'Dashboard',
+      menu: 'auth',
+      user,
+      userName,
+      playlists,
+      errors: err.details,
+      payload: request.payload,
+    })
+    .takeover();
+};
+
 export const addPlaylist = async (request, h) => {
   const user = request.auth.credentials;
   const { title } = request.payload;